Adding factory method on configuration to simplify explicit interface implementation
