+++ src/AutoMapper/MapperConfiguration.cs
@@ -11,22 +11,6 @@ namespace AutoMapper
     using QueryableExtensions;
     using QueryableExtensions.Impl;
 
-    public class TypeMapRegistry
-    {
-        private readonly ConcurrentDictionary<TypePair, TypeMap> _typeMaps = new ConcurrentDictionary<TypePair, TypeMap>();
-
-        public IEnumerable<TypeMap> TypeMaps => _typeMaps.Values;
-
-        public void RegisterTypeMap(TypeMap typeMap) => _typeMaps.AddOrUpdate(typeMap.Types, typeMap, (tp, tm) => tm);
-
-        public TypeMap GetTypeMap(TypePair typePair)
-        {
-            TypeMap typeMap;
-
-            return _typeMaps.TryGetValue(typePair, out typeMap) ? typeMap : null;
-        }
-    }
-
     public class MapperConfiguration : IConfigurationProvider, IMapperConfiguration
     {
         private readonly IEnumerable<IObjectMapper> _mappers;
@@ -259,6 +243,10 @@ namespace AutoMapper
             AssertConfigurationIsValid(_typeMapRegistry.TypeMaps);
         }
 
+        public IMapper CreateMapper() => new Mapper(this);
+
+        public IMapper CreateMapper(Func<Type, object> serviceCtor) => new Mapper(this, serviceCtor);
+
         public IEnumerable<IObjectMapper> GetMappers() => _mappers;
 
         public IEnumerable<ITypeMapObjectMapper> GetTypeMapMappers() => _typeMapObjectMappers;
@@ -321,10 +309,6 @@ namespace AutoMapper
             }
         }
 
-
-        public IMapper CreateMapper() => new Mapper(this);
-
-        public IMapper CreateMapper(Func<Type, object> serviceCtor) => new Mapper(this, serviceCtor);
 
         private IEnumerable<TypeMap> GetDerivedTypeMaps(TypeMap typeMap)
         {
